+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next'
 import './globals.css'
-import JotaiProvider from './_provider/JotaiProvider'
-import TopNavigation from './components/templates/topNavigation'
-import ReactQueryProvider from './_provider/ReactQueryProvider'
+import JotaiProvider from './providers/JotaiProvider'
+import TopNavigation from './components/templates/TopNavigation'
+import ReactQueryProvider from './providers/ReactQueryProvider'
 
 export const metadata: Metadata = {
   title: 'Nextjs15',
